+++ tests/bridge_balance_erc20_test.py
@@ -4,7 +4,8 @@ import unittest
 from substrateinterface import SubstrateInterface, Keypair
 from peaq.utils import ExtrinsicBatch
 from peaq.utils import get_account_balance
-from tools.utils import get_modified_chain_spec, sign_and_submit_evm_transaction
+from tools.utils import get_modified_chain_spec
+from tools.peaq_eth_utils import sign_and_submit_evm_transaction
 from tools.utils import WS_URL, ETH_URL
 from tools.utils import KP_GLOBAL_SUDO
 from tools.peaq_eth_utils import get_contract
