+++ tests/bridge_did_test.py
@@ -6,7 +6,7 @@ from peaq.eth import calculate_evm_account, calculate_evm_addr
 from peaq.extrinsic import transfer
 from tools.peaq_eth_utils import call_eth_transfer_a_lot, get_contract, generate_random_hex
 from tools.utils import WS_URL, ETH_URL
-from tools.utils import sign_and_submit_evm_transaction
+from tools.peaq_eth_utils import sign_and_submit_evm_transaction
 from tools.peaq_eth_utils import get_eth_chain_id
 from web3 import Web3
