+++ tests/bridge_xcmutils_test.py
@@ -12,7 +12,7 @@ from substrateinterface import SubstrateInterface, Keypair
 from peaq.utils import ExtrinsicBatch
 from web3 import Web3
 from tools.utils import KP_GLOBAL_SUDO
-from tools.utils import sign_and_submit_evm_transaction
+from tools.peaq_eth_utils import sign_and_submit_evm_transaction
 from peaq.utils import get_account_balance
 from tests import utils_func as TestUtils
 from tools.asset import PEAQ_ASSET_LOCATION, PEAQ_METADATA, PEAQ_ASSET_ID
