+++ tests/bridge_xtokens_test.py
@@ -8,7 +8,7 @@ from tools.runtime_upgrade import wait_until_block_height
 from substrateinterface import SubstrateInterface, Keypair
 from tools.utils import ETH_URL, RELAYCHAIN_WS_URL
 from tools.utils import WS_URL, ACA_WS_URL, PARACHAIN_WS_URL
-from tools.utils import sign_and_submit_evm_transaction
+from tools.peaq_eth_utils import sign_and_submit_evm_transaction
 from peaq.utils import get_account_balance
 from peaq.utils import ExtrinsicBatch
 from tools.asset import convert_enum_to_asset_id
