+++ tools/monkey/monkey_patch_scale_info.py
